LAR-12: Refactor Show Reminder Test Cases
- Adding test cases to validate show reminder
