+++ src/tests/Feature/Reminder/UserCanShowReminderTest.php
@@ -27,23 +27,4 @@ class UserCanShowReminderTest extends TestCase
                 'data' => $reminder->only('id', 'title', 'description', 'remind_at', 'event_at'),
             ]);
     }
-
-    /** @test */
-    public function cannot_show_reminder_by_id_because_resource_not_found(): void
-    {
-        Sanctum::actingAs($user = User::factory()->create(), ['access-api']);
-
-        $response = $this->getJson('/api/reminders/1');
-
-        $response
-            ->assertNotFound()
-            ->assertJson([
-                'ok' => false,
-                'data' => [
-                    'id' => [__('validation.exists', ['attribute' => 'id'])],
-                ],
-                'err' => 'ERR_NOT_FOUND',
-                'msg' => 'resource is not found',
-            ]);
-    }
 }
